crypto-stream-service/requirements.txt
boto3==1.24.17
botocore==1.27.17
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.0.12
cryptowatch-sdk==0.0.16
idna==2.10
jmespath==1.0.1
marshmallow==3.13.0
pika==1.2.1
protobuf==3.20.1
python-dateutil==2.8.2
PyYAML==6.0
requests==2.25.1
s3transfer==0.6.0
schedule==1.1.0
six==1.16.0
urllib3==1.26.9
websocket-client==1.3.3
